ref(compiler): split three multi-responsibility classes into collaborators
Mirrors the CallSpecialization split. Pure refactor, behavior unchanged.

- DefSymbol: extract MacroFormRewriter (macro &form/&env injection + ^:tag
  return-type form rewriting, unifying the single-/multi-arity walkers) and
  DefArglistBuilder (:arglists string formatting).
- ParamTypeInferrer: extract CallTypeExpectationResolver (stateless PHP-call
  and phel.core-call shape analysis); the walker and its mutable observation
  state stay in the inferrer.
- TestCommand: extract TestCommandOptionParser (option parsing/coercion +
  parallelism decision); it owns the CLI option-name constants.

Related to #2262
